Working bar for the landing page info section
