+++ app/client/style.css
@@ -15,4 +15,8 @@
   left:30%
 }
 
-
+.background-bar {
+  background-size: cover;
+  background: #17252a no-repeat center center fixed;
+  height: 450px;
+}
