+++ app/imports/ui/components/FullWidthImage.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Header, Container, Grid, Icon, Image } from 'semantic-ui-react';
+import LandingBar from './LandingBar';
 
 export default class FullWidthImage extends React.Component {
   render() {
@@ -24,17 +25,12 @@ export default class FullWidthImage extends React.Component {
       borderRadius: '20px',
       paddingBottom: '16px',
       paddingTop: '16px',
+      marginBottom: '360px',
     };
     const headerStyle = {
       fontFamily: 'Baven',
       opacity: '10 !important',
       color: '#17252a',
-    };
-    const headerStyle2 = {
-      marginTop: '600px',
-      fontFamily: 'Baven',
-      paddingBottom: '16px',
-      paddingTop: '16px',
     };
     const imageStyle = { opacity: '0.9' };
     return (
@@ -55,20 +51,9 @@ export default class FullWidthImage extends React.Component {
                 </Header>
               </Container>
 
-              <Container fluid style={infoContainerStyle}>
-                <Header style={headerStyle2} as='h2' icon textAlign='center'>
-                  <Icon name='users' circular/>
-                  <Header.Content>A community for the community</Header.Content>
-                </Header>
-                <Header style={headerStyle} textAlign='center' as='h3'>
-                  At UH Bazaar, the focus is the student.  What UHB aims to provide is
-                  a comfortable and easy experience when it comes to getting the things
-                  you need.  Leaving more time and energy to focus on what matters: Learning.
-                </Header>
-              </Container>
-
             </Container>
           </Grid>
+          <LandingBar/>
         </Container>
     );
   }
